finalize subsribe module
